Add files via upload
The Blood Bank System is a web-based application developed using HTML, Python, and Django, designed to manage blood donations and requests effectively. It provides a platform for donors to offer blood and for recipients to request blood as needed. The system is designed to enhance the efficiency of blood management and accessibility.

Key Features:
• User Registration & Authentication: Secure user registration and login for donors and recipients.
• Blood Donation Management: Donors can register their details and donate blood, with donation history tracked.
• Blood Type Availability Search: Users can search for available blood types in real-time based on location and urgency.
• Admin Dashboard: Admins can manage donor information, track donations, and monitor blood inventory.
• Request Management: Users can request blood, and admins can approve or reject requests based on availability.
• Notifications: Automated notifications are sent to donors for blood donation reminders and recipients for blood request updates.
• Reporting: Admins can generate reports on donation statistics, inventory levels, and donor participation.
• Data Security: Secure data handling with encrypted transactions and compliance with privacy regulations.
The system bridges the gap between blood donors and recipients, ensuring a steady and accessible blood supply for hospitals and patients in need.